+++ api/lib/domain/usecases/index.js
@@ -48,7 +48,6 @@ import * as assessmentRepository from '../../../src/shared/infrastructure/reposi
 import * as challengeRepository from '../../../src/shared/infrastructure/repositories/challenge-repository.js';
 import * as competenceRepository from '../../../src/shared/infrastructure/repositories/competence-repository.js';
 import * as courseRepository from '../../../src/shared/infrastructure/repositories/course-repository.js';
-import * as frameworkRepository from '../../../src/shared/infrastructure/repositories/framework-repository.js';
 import { repositories as sharedRepositories } from '../../../src/shared/infrastructure/repositories/index.js';
 import * as organizationLearnerRepository from '../../../src/shared/infrastructure/repositories/organization-learner-repository.js';
 import * as organizationRepository from '../../../src/shared/infrastructure/repositories/organization-repository.js';
@@ -119,7 +118,6 @@ const dependencies = {
   dataProtectionOfficerRepository,
   emailRepository,
   emailValidationDemandRepository,
-  frameworkRepository,
   userToCreateRepository,
   knowledgeElementRepository: sharedRepositories.knowledgeElementRepository,
   lastUserApplicationConnectionsRepository,
