+++ api/src/learning-content/domain/usecases/get-framework-areas.js
@@ -1,4 +1,4 @@
-import { AreaForAdmin } from '../../../src/shared/domain/models/index.js';
+import { AreaForAdmin } from '../../../shared/domain/models/index.js';
 
 const getFrameworkAreas = async function ({
   frameworkId,
