+++ api/src/learning-content/infrastructure/serializers/framework-areas-serializer.js
