+++ api/src/learning-content/routes.js
@@ -1,3 +1,4 @@
+import * as frameworksRoute from './application/frameworks-route.js';
 import * as learningContentRoute from './application/learning-content-route.js';
 
-export const learningContentRoutes = [learningContentRoute];
+export const learningContentRoutes = [learningContentRoute, frameworksRoute];
