+++ api/tests/learning-content/unit/application/frameworks-controller_test.js
@@ -1,5 +1,5 @@
-import { frameworksController } from '../../../../lib/application/frameworks/frameworks-controller.js';
-import { usecases } from '../../../../lib/domain/usecases/index.js';
+import { frameworksController } from '../../../../src/learning-content/application/frameworks-controller.js';
+import { usecases } from '../../../../src/learning-content/domain/usecases/index.js';
 import { expect, hFake, sinon } from '../../../test-helper.js';
 
 describe('Unit | Controller | frameworks-controller', function () {
